requirements.txt
# Core dependencies
torch>=1.9.0
numpy>=1.19.0

# Neural ODE and differential equations
torchdiffeq>=0.2.0
torchcde>=0.2.0

# Optimal transport and geometry
geomloss>=0.2.0
POT>=0.8.0

# Scientific computing
scipy>=1.5.0
pandas>=1.1.0
scikit-learn>=0.24.0
scikit-optimize>=0.8.0

# Visualization
matplotlib>=3.3.0
seaborn>=0.11.0
imageio>=2.9.0

# Single-cell analysis
scanpy>=1.7.0
anndata>=0.7.0
harmonypy>=0.0.5

# Configuration
pyyaml>=5.4.0

# Utilities
joblib>=1.0.0
tqdm>=4.50.0
